+++ crates/common/src/platform.rs
@@ -19,6 +19,7 @@ use wasefire_error::{Code, Error};
 /// Platform side.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub enum Side {
     A,
     B,