+++ crates/error/test.sh
@@ -21,5 +21,6 @@ test_helper
 
 cargo test --lib --features=std
 cargo check --lib
+cargo check --lib --features=serde
 cargo check --lib --target=thumbv7em-none-eabi --features=defmt
 cargo check --lib --target=wasm32-unknown-unknown